removed legacy metrics
