+++ lib/metrics.js
@@ -33,10 +33,8 @@ var Metrics = function(opts) {
 	this.opts = {};
 	this.httpReq		= new Proxies.HttpRequest();
 	this.httpRes		= new Proxies.HttpResponse();
-	this.httpResV2		= new Proxies.HttpResponseV2();
 	this.system		 = new Proxies.System();
 	this.fetch		  = new Proxies.Fetch();
-	this.ftApiClient	= new Proxies.FtApiClient();
 	this.aggregators = [];
 
 	// arbitrary counters
@@ -125,11 +123,7 @@ Metrics.prototype.flushRate = function () {
 Metrics.prototype.setupDefaultAggregators = function () {
 	this.registerAggregator(this.flushCustomCounters.bind(this));
 	this.registerAggregator(this.flushCustomHistograms.bind(this));
-	this.registerAggregator(this.httpReq.reporter);
-	this.registerAggregator(this.httpRes.reporter);
-	this.registerAggregator(this.httpResV2.reporter);
 	this.registerAggregator(this.system.counts);
-	this.registerAggregator(this.ftApiClient.reporter);
 	this.registerAggregator(this.flushRate.bind(this));
 };
 
@@ -146,13 +140,11 @@ Metrics.prototype.instrument = function (obj, opts) {
 	switch (opts.as) {
 		case 'express.http.res':
 			this.httpRes.instrument(obj);
-			this.httpResV2.instrument(obj);
+			this.registerAggregator(this.httpRes.reporter);
 			break;
 		case 'express.http.req':
 			this.httpReq.instrument(obj);
-			break;
-		case 'ft-api-client':
-			this.ftApiClient.instrument(obj);
+			this.registerAggregator(this.httpReq.reporter);
 			break;
 		default:
 			throw new Error('No valid "opts.as" argument given. You need to specify the object you want instrumented.');
