+++ lib/metrics/express/http-response.js
@@ -1,37 +1,28 @@
 "use strict";
 
-var _ = require('lodash');
 var metrics = require('metrics');
 
 var HttpResponse = module.exports = function HttpResponse() {
-	this.counters = {
-		status_2xx:  new metrics.Counter(),
-		status_3xx:  new metrics.Counter(),
-		status_4xx:  new metrics.Counter(),
-		status_5xx:  new metrics.Counter(),
-		status_2xx_response_time: new metrics.Histogram.createUniformHistogram(),
-		status_4xx_response_time: new metrics.Histogram.createUniformHistogram(),
-		status_5xx_response_time: new metrics.Histogram.createUniformHistogram(),
-	};
-
-	this.dynamicCounters = ['template_render_time'];
-
-	var value = function() {
-		_.invoke(this, 'clear');
-	};
-
-	for (var i= 0,l=this.dynamicCounters.length; i<l; i++) {
-		this.counters[this.dynamicCounters[i]] = {};
-		Object.defineProperty(this.counters[this.dynamicCounters[i]], 'clear', {
-			value: value
-		});
-	}
+	this.buckets = {};
 	this.reporter = this.reporter.bind(this);
 };
 
+HttpResponse.prototype.measure = function (name, value) {
+	var bucket = this.buckets[name] || (this.buckets[name] = {
+		counter: new metrics.Counter(),
+		histogram: new metrics.Histogram.createUniformHistogram()
+	});
+
+	bucket.counter.inc(1);
+	if (value) {
+		bucket.histogram.update(value);
+	}
+};
+
 HttpResponse.prototype.instrument = function(res) {
+
 	if (!res._nextInstrumented) {
-		res.writeHead = this._writeHead(res.writeHead, Date.now());
+		res.writeHead = this._writeHead(res.writeHead, res.req.method, Date.now());
 		res.render = (function() {
 			var _render = res.render.bind(res),
 				httpResponse = this;
@@ -41,11 +32,7 @@ HttpResponse.prototype.instrument = function(res) {
 						returnValue = _render(template, data, callback),
 						diff = process.hrtime(start);
 
-					if(!httpResponse.counters.template_render_time[template]) {
-						httpResponse.counters.template_render_time[template] = new metrics.Histogram.createUniformHistogram();
-					}
-
-					httpResponse.counters.template_render_time[template].update((diff[0] * 1e9) + diff[1]);
+					httpResponse.measure('express.default_route_' + res.req.method + '.res.template_render.' + template, (diff[0] * 1e9) + diff[1]);
 					return returnValue;
 				};
 
@@ -55,77 +42,26 @@ HttpResponse.prototype.instrument = function(res) {
 };
 
 HttpResponse.prototype.reset = function() {
-	_.invoke(this.counters, 'clear');
+	Object.keys(this.buckets).forEach(function (key) {
+		this.buckets[key].counter.clear();
+		this.buckets[key].histogram.clear();
+	}.bind(this));
 };
 
+
 HttpResponse.prototype.counts = function() {
-	var c = this.counters;
-	var obj = _.zipObject(
-		[
-			'express.http.res.status_2xx_response_time.mean',
-			'express.http.res.status_2xx_response_time.stdDev',
-			'express.http.res.status_2xx_response_time.min',
-			'express.http.res.status_2xx_response_time.max',
-			'express.http.res.status_2xx_response_time.sum',
-			'express.http.res.status_4xx_response_time.mean',
-			'express.http.res.status_4xx_response_time.stdDev',
-			'express.http.res.status_4xx_response_time.min',
-			'express.http.res.status_4xx_response_time.max',
-			'express.http.res.status_4xx_response_time.sum',
-			'express.http.res.status_5xx_response_time.mean',
-			'express.http.res.status_5xx_response_time.stdDev',
-			'express.http.res.status_5xx_response_time.min',
-			'express.http.res.status_5xx_response_time.max',
-			'express.http.res.status_5xx_response_time.sum',
-			'express.http.res.status_2xx.count',
-			'express.http.res.status_3xx.count',
-			'express.http.res.status_4xx.count',
-			'express.http.res.status_5xx.count',
-			'express.http.res.render.min',
-			'express.http.res.render.max',
-			'express.http.res.render.sum',
-			'express.http.res.render.mean',
-			'express.http.res.render.count'
-		],
-		[
-			c.status_2xx_response_time.mean(),
-			c.status_2xx_response_time.stdDev(),
-			c.status_2xx_response_time.min,
-			c.status_2xx_response_time.max,
-			c.status_2xx_response_time.sum,
-			c.status_4xx_response_time.mean(),
-			c.status_4xx_response_time.stdDev(),
-			c.status_4xx_response_time.min,
-			c.status_4xx_response_time.max,
-			c.status_4xx_response_time.sum,
-			c.status_5xx_response_time.mean(),
-			c.status_5xx_response_time.stdDev(),
-			c.status_5xx_response_time.min,
-			c.status_5xx_response_time.max,
-			c.status_5xx_response_time.sum,
-			c.status_2xx.count,
-			c.status_3xx.count,
-			c.status_4xx.count,
-			c.status_5xx.count
-		]
-	);
-	// replace characters that will look odd in metrics with underscore
-	function replaceOddCharactersWithUnderscore(template) {
-		var safeName = template.replace(/[\/\\\.]/g, '_');
-		var str = 'express.http.res.render.'+safeName+'.';
-		obj[str + 'min'] = c.template_render_time[template].min;
-		obj[str + 'max'] = c.template_render_time[template].max;
-		obj[str + 'sum'] = c.template_render_time[template].sum;
-		obj[str + 'mean'] = c.template_render_time[template].mean();
-		obj[str + 'count'] = c.template_render_time[template].count;
-	}
+	var metrics = {};
+	Object.keys(this.buckets).forEach(function (key) {
+		var safeKey = (key.indexOf('template' > -1) ? key.replace(/(!?:template_render)[\/\\\.]/g, '_') : key);
+		metrics[safeKey + '.count'] = this.buckets[key].counter.count;
+		metrics[safeKey + '.time.sum'] = this.buckets[key].histogram.sum;
+		metrics[safeKey + '.time.max'] = this.buckets[key].histogram.max;
+		metrics[safeKey + '.time.min'] = this.buckets[key].histogram.min;
+		metrics[safeKey + '.time.mean'] = this.buckets[key].histogram.mean();
+		metrics[safeKey + '.time.stdDev'] = this.buckets[key].histogram.stdDev();
+	}.bind(this));
 
-	for (var i= 0, l=this.dynamicCounters.length; i<l; i++) {
-		var counter = c[this.dynamicCounters[i]];
-		Object.keys(counter).forEach(replaceOddCharactersWithUnderscore);
-	}
-
-	return obj;
+	return metrics;
 };
 
 HttpResponse.prototype.reporter = function() {
@@ -136,28 +72,11 @@ HttpResponse.prototype.reporter = function() {
 
 
 // proxy for res.writeHead - http://nodejs.org/api/http.html#http_response_writehead_statuscode_reasonphrase_headers
-HttpResponse.prototype._writeHead = function(fn, dt) {
+HttpResponse.prototype._writeHead = function(fn, method, dt) {
 	var self = this;
-	return function() {
-		var statusCode = parseInt(('' + arguments[0]).toString().charAt(0));
-		switch (statusCode) {
-			case 2:
-				self.counters.status_2xx.inc(1);
-				self.counters.status_2xx_response_time.update(Date.now() - dt);
-				break;
-			case 3:
-				self.counters.status_3xx.inc(1);
-				break;
-			case 4:
-				self.counters.status_4xx.inc(1);
-				self.counters.status_4xx_response_time.update(Date.now() - dt);
-				break;
-			case 5:
-				self.counters.status_5xx.inc(1);
-				self.counters.status_5xx_response_time.update(Date.now() - dt);
-				break;
-			default:
-		}
-		fn.apply(this, arguments);
+	return function () {
+		var statusCode = parseInt(('' + arguments[0]).toString().substr(0, 3));
+		self.measure('express.default_route_' + method + '.res.status.' + statusCode, Date.now() - dt);
+		return fn.apply(this, arguments);
 	};
 };
