+++ lib/metrics/index.js
@@ -1,6 +1,4 @@
 exports.HttpRequest     = require('./express/http-request');
 exports.HttpResponse    = require('./express/http-response');
-exports.HttpResponseV2    = require('./express/http-response-v2');
 exports.System          = require('./system/process');
 exports.Fetch           = require('./fetch');
-exports.FtApiClient     = require('./ft-api-client');
